+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 - ✨(frontend) leave a document #2410
 - ✨(frontend) add top parent on sub docs search #1952
 - ✨(frontend) unauthenticated users can search #2407
+- ✨(backend) specific user delete method to delete its relations
 
 ### Changed